+++ cmd/gocritic/testdata/unusedParam/negative_tests.go
